+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Ignore received replication for entities that was despawned by the client.
+
 ## [0.37.0] - 2025-12-03
 
 ### Added